Move ringDescriber to a separate file and make in not rely on Session
