+++ ring_describer.go
@@ -1,0 +1,117 @@
+package gocql
+
+import (
+	"context"
+	"fmt"
+	"sync"
+)
+
+// Polls system.peers at a specific interval to find new hosts
+type ringDescriber struct {
+	control         *controlConn
+	cfg             *ClusterConfig
+	logger          StdLogger
+	mu              sync.Mutex
+	prevHosts       []*HostInfo
+	prevPartitioner string
+}
+
+func (r *ringDescriber) setControlConn(c *controlConn) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+
+	r.control = c
+}
+
+// Ask the control node for the local host info
+func (r *ringDescriber) getLocalHostInfo() (*HostInfo, error) {
+	if r.control == nil {
+		return nil, errNoControl
+	}
+
+	iter := r.control.withConnHost(func(ch *connHost) *Iter {
+		return ch.conn.querySystemLocal(context.TODO())
+	})
+
+	if iter == nil {
+		return nil, errNoControl
+	}
+
+	host, err := hostInfoFromIter(iter, nil, r.cfg.Port, r.cfg.translateAddressPort)
+	if err != nil {
+		return nil, fmt.Errorf("could not retrieve local host info: %w", err)
+	}
+	return host, nil
+}
+
+// Ask the control node for host info on all it's known peers
+func (r *ringDescriber) getClusterPeerInfo(localHost *HostInfo) ([]*HostInfo, error) {
+	if r.control == nil {
+		return nil, errNoControl
+	}
+
+	var peers []*HostInfo
+	iter := r.control.withConnHost(func(ch *connHost) *Iter {
+		return ch.conn.querySystemPeers(context.TODO(), localHost.version)
+	})
+
+	if iter == nil {
+		return nil, errNoControl
+	}
+
+	rows, err := iter.SliceMap()
+	if err != nil {
+		// TODO(zariel): make typed error
+		return nil, fmt.Errorf("unable to fetch peer host info: %s", err)
+	}
+
+	for _, row := range rows {
+		// extract all available info about the peer
+		host, err := hostInfoFromMap(row, &HostInfo{port: r.cfg.Port}, r.cfg.translateAddressPort)
+		if err != nil {
+			return nil, err
+		} else if !isValidPeer(host) {
+			// If it's not a valid peer
+			r.logger.Printf("Found invalid peer '%s' "+
+				"Likely due to a gossip or snitch issue, this host will be ignored", host)
+			continue
+		}
+
+		peers = append(peers, host)
+	}
+
+	return peers, nil
+}
+
+// Return true if the host is a valid peer
+func isValidPeer(host *HostInfo) bool {
+	return !(len(host.RPCAddress()) == 0 ||
+		host.hostId == "" ||
+		host.dataCenter == "" ||
+		host.rack == "" ||
+		len(host.tokens) == 0)
+}
+
+// GetHosts returns a list of hosts found via queries to system.local and system.peers
+func (r *ringDescriber) GetHosts() ([]*HostInfo, string, error) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+
+	localHost, err := r.getLocalHostInfo()
+	if err != nil {
+		return r.prevHosts, r.prevPartitioner, err
+	}
+
+	peerHosts, err := r.getClusterPeerInfo(localHost)
+	if err != nil {
+		return r.prevHosts, r.prevPartitioner, err
+	}
+
+	hosts := append([]*HostInfo{localHost}, peerHosts...)
+	var partitioner string
+	if len(hosts) > 0 {
+		partitioner = hosts[0].Partitioner()
+	}
+
+	return hosts, partitioner, nil
+}
